pyproject.toml
# SPDX-FileCopyrightText: 2023-2024 Steffen Vogel, OPAL-RT Germany GmbH
# SPDX-License-Identifier: Apache-2.0

[tool.poetry]
authors = [
	"Steffen Vogel <steffen.vogel@opal-rt.com>",
	"Felix Wege <felix.wege@eonerc.rwth-aachen.de>",
]
name = "seguro-platform"
version = "0.1.0"
description = "Common Python helpers for accessing the SEGuRo platform"
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "seguro" }]

[tool.poetry.scripts]
poetry = "poetry.console.application:main"

scheduler = "seguro.commands.scheduler.main:main"
recorder = "seguro.commands.recorder.main:main"
notifier = "seguro.commands.notifier.main:main"
signature-sender = "seguro.commands.signature_sender.main:main"
heartbeat-sender = "seguro.commands.heartbeat_sender.main:main"
signature-recorder = "seguro.commands.signature_recorder.main:main"
notebook-executor = "seguro.commands.notebook_executor.main:main"
acl-syncer = "seguro.commands.acl_syncer.main:main"
demo-data = "seguro.commands.demo_data.main:main"
protobuf-subscriber = "seguro.commands.protobuf_subscriber.main:main"
fiware-connector = "seguro.commands.fiware_connector.main:main"
flex-marketing-service = "seguro.commands.flex_marketing_service.main:main"

[tool.poetry.urls]
"Homepage" = "https://github.com/SEGuRo-Projekt/Platform"
"Bug Tracker" = "https://github.com/SEGuRo-Projekt/Platform/issues"

[tool.black]
line-length = 79

[tool.pytest.ini_options]
markers = ["broker", "store", "logger", "config"]

[tool.poetry.dependencies]
python = "^3.11"
apprise = "^1.7.2"
boto3 = "^1.34.50"
django-environ = "^0.11.2"
docker = "^7.0.0"
minio = "^7.2.4"
paho-mqtt = "^2.0.0"
python-slugify = "^8.0.4"
pyyaml = "^6.0.1"
rfc3161ng = "^2.1.3"
schedule = "^1.2.1"
pandas = { extras = ["excel", "parquet"], version = "^2.2.1" }
cryptography = "42.0.3"
aws-logging-handlers = "^2.0.5"
s3fs = "^2024.3.1"
fsspec = "^2024.3.1"
pyusb = "^1.2.1"
psutil = "^5.9.8"
villas-node = "0.11.1"
pydantic = "^2.7.0"
pyarrow = "^15"
pytimeparse = "^1.1.8"
requests = ">=2.31.0"
pyvolt = { git = "https://github.com/sogno-platform/pyvolt" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
ipython = "^8.22.1"
flake8 = "^7.0.0"
black = "^24.2.0"
mypy = "^1.9.0"
pre-commit = "^3.6.2"
types-pyyaml = "^6.0.12.20240311"
types-python-slugify = "^8.0.2.20240310"
pandas-stubs = "^2.2.1.240316"
datamodel-code-generator = "^0.25.5"

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-copybutton = "^0.5.2"
sphinx-autobuild = "^2024.2.4"
sphinx-inline-tabs = "^2023.4.21"
furo = "^2024.1.29"
myst-parser = "^2.0.0"
sphinxext-opengraph = "^0.9.1"
sphinxcontrib-bibtex = "^2.6.2"
autodoc-pydantic = "^2.1.0"
erdantic = "^1.0.2"


[tool.poetry.group.jupyter.dependencies]
jupyterlab = "^4.1.5"
ipykernel = "^6.29.4"
nbconvert = "^7.16.3"

[tool.mypy]
plugins = ["pydantic.mypy"]

ignore_missing_imports = true
check_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.datamodel-codegen]
output-model-type = "pydantic_v2.BaseModel"
use-standard-collections = true
use-union-operator = true
target-python-version = "3.11"
enum-field-as-literal = "one"
capitalise-enum-members = true
use-field-description = true
use-schema-description = true
use-double-quotes = true
field-constraints = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
